separate the index to MVC
