+++ static/index.html
@@ -162,5 +162,5 @@
   </body>
   <script type="module" src="/static/js/auth.js"></script>
   <script type="module" src="/static/js/booking.js"></script>
-  <script src="/static/js/main.js"></script>
+  <script type="module" src="/static/js/main.js"></script>
 </html>
