+++ static/js/auth.js
@@ -1,12 +1,8 @@
 import { initialize, checkUserState } from "./controller/controller_auth.js";
-
-const signinMask = document.querySelector(".signin-mask");
-const signupMask = document.querySelector(".signup-mask");
-
-signinMask.style.display = "none";
-signupMask.style.display = "none";
+import { signUpSignInDisplayNone } from "./view/view.js";
 
 document.addEventListener("DOMContentLoaded", function () {
+  signUpSignInDisplayNone();
   initialize();
   checkUserState();
 });
