+++ static/js/main.js
@@ -1,9 +1,8 @@
+import * as IndexView from "./view/view_index.js";
+
 const leftContainerBtn = document.querySelector(".left-container");
 const rightContainerBtn = document.querySelector(".right-container");
 const scrollableContainer = document.getElementById("scrollable-container");
-
-const signinMask = document.querySelector(".signin-mask");
-const signupMask = document.querySelector(".signup-mask");
 
 const searchInput = document.querySelector(".searchKeyword");
 const searchButton = document.querySelector(".input-area button");
@@ -14,13 +13,6 @@ const attractionURL = "/api/attractions";
 let currentPage = 0;
 let hasNextPage = true;
 let isWaitingForData = false;
-
-let isDown = false;
-let startX;
-let scrollLeft;
-
-signinMask.style.display = "none";
-signupMask.style.display = "none";
 
 document.addEventListener("DOMContentLoaded", initializePage);
 
@@ -33,15 +25,94 @@ function initializePage() {
 
 // 各種功能性的函數呼叫
 function setupEventListeners() {
-  leftContainerBtn.addEventListener("click", leftScroll);
-  rightContainerBtn.addEventListener("click", rightScroll);
-  searchButton.addEventListener("click", search);
+  leftContainerBtn.addEventListener("click", IndexView.leftScroll);
+  rightContainerBtn.addEventListener("click", IndexView.rightScroll);
+  searchButton.addEventListener("click", searchEvent);
   searchInput.addEventListener("keypress", enterPress);
 
-  scrollableContainer.addEventListener("mousedown", mousedown);
-  scrollableContainer.addEventListener("mouseleave", mouseleave);
-  scrollableContainer.addEventListener("mouseup", mouseup);
-  scrollableContainer.addEventListener("mousemove", mousemove);
+  scrollableContainer.addEventListener("mousedown", IndexView.mousedown);
+  scrollableContainer.addEventListener("mouseleave", IndexView.mouseleave);
+  scrollableContainer.addEventListener("mouseup", IndexView.mouseup);
+  scrollableContainer.addEventListener("mousemove", IndexView.mousemove);
+}
+
+function enterPress(event) {
+  if (event.key === "Enter") {
+    event.preventDefault();
+    searchEvent(event);
+  }
+}
+
+function searchEvent(event) {
+  event.preventDefault();
+
+  if (isWaitingForData) return;
+
+  refresh();
+
+  const searchInput = document.querySelector(".searchKeyword");
+  const keywordInputValue = searchInput.value;
+
+  fetchGetAttractions(keywordInputValue, currentPage, true).then(() => {});
+  console.log(keywordInputValue);
+}
+
+function refresh() {
+  currentPage = 0;
+  IndexView.refreshContent();
+}
+
+const observer = new IntersectionObserver(
+  (entries) => {
+    const firstEntry = entries[0];
+
+    if (firstEntry.isIntersecting && hasNextPage && !isWaitingForData) {
+      const keywordInputValue = searchInput.value;
+      //調用fetch函式的時候使用非同步加載
+      fetchGetAttractions(keywordInputValue, currentPage + 1).then(() => {});
+    }
+  },
+  { threshold: 0.5 }
+);
+
+//fetch GET API頁面的景點
+async function fetchGetAttractions(keyword = "") {
+  try {
+    //開始新的資料加載前設定
+    isWaitingForData = true;
+
+    const response = await fetch(
+      `${attractionURL}?keyword=${keyword}&page=${currentPage}`
+    );
+    if (!response.ok) {
+      throw new Error(`Server responded with status ${response.status}`);
+    }
+
+    const data = await response.json();
+    if (!data || !Array.isArray(data.data)) {
+      throw new Error("Invalid data structure");
+    }
+
+    currentPage++;
+    hasNextPage = data.nextPage != null;
+    console.log("currentPage :", currentPage);
+
+    let lastItem = document.querySelector(".grid-item:last-child");
+
+    IndexView.displayAttractions(data.data, keyword);
+
+    let newItem = document.querySelector(".grid-item:last-child");
+
+    if (lastItem) observer.unobserve(lastItem);
+    if (hasNextPage) {
+      if (newItem) observer.observe(newItem);
+    }
+
+    isWaitingForData = false;
+  } catch (error) {
+    console.error("Error fetching attractions:", error);
+    isWaitingForData = false;
+  }
 }
 
 // 頁面初始化的載入API的MRT
@@ -69,158 +140,4 @@ async function fetchGetMRTStations() {
   } catch (error) {
     console.error("Error fetching MRT stations:", error);
   }
-}
-
-//fetch GET API頁面的景點
-async function fetchGetAttractions(keyword = "", page = 0, refresh = false) {
-  try {
-    //開始新的資料加載前設定
-    isWaitingForData = true;
-    const response = await fetch(
-      `${attractionURL}?keyword=${keyword}&page=${page}`
-    );
-    if (!response.ok) {
-      throw new Error(`Server responded with status ${response.status}`);
-    }
-
-    const data = await response.json();
-    if (!data || !Array.isArray(data.data)) {
-      throw new Error("Invalid data structure");
-    }
-    displayAttractions(data.data, keyword, refresh);
-    currentPage = page;
-    hasNextPage = data.nextPage != null;
-    isWaitingForData = false;
-  } catch (error) {
-    console.error("Error fetching attractions:", error);
-    isWaitingForData = false;
-  }
-}
-
-function displayAttractions(attractions, keyword, refresh = false) {
-  lastItem = document.querySelector(".grid-item:last-child");
-
-  const attractionsContainer = document.querySelector(".attractions-group");
-  if (refresh) {
-    currentPage == 0;
-    attractionsContainer.innerHTML = "";
-  }
-  if (attractions.length === 0) {
-    attractionsContainer.innerHTML = "";
-    const noDataMessage = document.createElement("div");
-    noDataMessage.textContent = `查無此"${keyword}"相關景點`;
-    noDataMessage.style.color = "red";
-    attractionsContainer.appendChild(noDataMessage);
-  } else {
-    attractions.forEach((attraction) => {
-      const gridItem = document.createElement("div");
-      gridItem.className = "grid-item";
-      gridItem.innerHTML = `
-    <div class="img">
-        <img src="${attraction.image[0]}" alt="${attraction.description}" />
-        <div class="location-name">
-            <p>${attraction.name}</p>
-        </div>
-    </div>
-    <div class="mrt-category">
-        <p>${attraction.mrt}</p>
-        <p>${attraction.category}</p>
-    </div>
-`;
-      gridItem.addEventListener("click", () => {
-        window.location.href = `/attraction/${attraction.id}`;
-      });
-      attractionsContainer.appendChild(gridItem);
-    });
-  }
-
-  newItem = document.querySelector(".grid-item:last-child");
-
-  if (lastItem) observer.unobserve(lastItem);
-  if (hasNextPage) {
-    if (newItem) observer.observe(newItem);
-  }
-}
-
-const observer = new IntersectionObserver(
-  (entries) => {
-    const firstEntry = entries[0];
-
-    if (firstEntry.isIntersecting && hasNextPage && !isWaitingForData) {
-      const keywordInputValue = searchInput.value;
-      //調用fetch函式的時候使用非同步加載
-      fetchGetAttractions(keywordInputValue, currentPage + 1).then(() => {});
-    }
-  },
-  { threshold: 0.5 }
-);
-
-function search(event) {
-  event.preventDefault();
-  if (isWaitingForData) return;
-
-  isWaitingForData = true;
-  currentPage = 0;
-
-  const keywordInputValue = searchInput.value;
-
-  fetchGetAttractions(keywordInputValue, currentPage, true).then(() => {});
-  console.log(keywordInputValue);
-}
-
-function enterPress(event) {
-  if (event.key === "Enter") {
-    event.preventDefault();
-    search(event);
-  }
-}
-
-function leftScroll(event) {
-  event.preventDefault();
-  const scrollDistance = getSrollDistance();
-  scrollableContainer.scrollLeft -= scrollDistance;
-  console.log("Scrolled left to:", scrollableContainer.scrollLeft);
-}
-
-function rightScroll(event) {
-  event.preventDefault();
-  const scrollDistance = getSrollDistance();
-  scrollableContainer.scrollLeft += scrollDistance;
-  console.log("Scrolled right to:", scrollableContainer.scrollLeft);
-}
-
-function getSrollDistance() {
-  const width = window.innerWidth;
-  if (width >= 601 && width <= 1200) {
-    return 200;
-  } else if (width >= 360 && width <= 600) {
-    return 100;
-  } else {
-    return 400;
-  }
-}
-
-function mousedown(event) {
-  isDown = true;
-  startX = event.pageX - scrollableContainer.offsetLeft;
-  scrollLeft = scrollableContainer.scrollLeft;
-  scrollableContainer.style.cursor = "grabbing";
-}
-
-function mouseleave() {
-  isDown = false;
-  scrollableContainer.style.cursor = "grab";
-}
-
-function mouseup() {
-  isDown = false;
-  scrollableContainer.style.cursor = "grab";
-}
-
-function mousemove(event) {
-  if (!isDown) return;
-  event.preventDefault;
-  const x = event.pageX - scrollableContainer.offsetLeft;
-  const speed = (x - startX) * 1.5;
-  scrollableContainer.scrollLeft = scrollLeft - speed;
 }
